renderer: Split into new subpackage with Orthogonal support only
We incorrectly try to render all map orientations with our current
map renderer, even though it only supports (the most common) orthogonal
orientation.

We'll roll renderers for each map orientation to avoid breaking the
existing implementations with dodgy conditionals, allowing each one
to be the most optimised implementation it can be.

This helps pave the way for isometric layout support.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>